add index.js and outsource function
